consolidate fishing and hunting shop commands into single entry points
Each game now has one shop command covering everything that costs coins:
- /fishshop: rods, upgrades, bait, consumables, repair, location unlock
- /huntshop: weapons, upgrades, ammo, consumables, repair, zone unlock

Removed: /fishbuyrod, /huntbuygun, /fishrepair, /huntrepair (all absorbed)
Updated: /fishlocation and /huntzone now navigation-only (unlock moved to shop)
Updated: all cross-file references to old command names
