+++ src/commands/economy/hunt.js
@@ -33,7 +33,7 @@ module.exports = {
         .setDescription('Hunt animals in your zone. Uses 1 stamina. Cooldown: 30s. Equip a weapon with /huntinv.')
         .addStringOption(o =>
             o.setName('zone')
-                .setDescription('Zone to hunt in (defaults to your active zone). Unlock more zones with /huntzone unlock.')
+                .setDescription('Zone to hunt in (defaults to your active zone). Unlock more zones with /huntshop unlock.')
                 .setRequired(false)
                 .addChoices(...ZONE_CHOICES)),
 
@@ -74,7 +74,7 @@ module.exports = {
         }
         if (!h.unlockedZones.includes(zoneId)) {
             return interaction.reply({
-                content: `You haven't unlocked **${zone.name}** yet. Use \`/huntzone unlock ${zoneId}\` to unlock it.`,
+                content: `You haven't unlocked **${zone.name}** yet. Use \`/huntshop unlock\` to unlock it.`,
                 ephemeral: true
             });
         }
@@ -115,7 +115,7 @@ module.exports = {
         // ── Weapon check ───────────────────────────────────────────────────
         if (h.equippedWeaponIndex < 0 || !h.weapons[h.equippedWeaponIndex]) {
             return interaction.reply({
-                content: `You don't have a weapon equipped! Buy one with \`/huntbuygun\` and equip it with \`/huntinv equip 1\`.`,
+                content: `You don't have a weapon equipped! Buy one with \`/huntshop weapon\` and equip it with \`/huntinv equip 1\`.`,
                 ephemeral: true
             });
         }
@@ -124,7 +124,7 @@ module.exports = {
 
         if (weapon.status === 'broken' || weapon.currentDurability <= 0) {
             return interaction.reply({
-                content: `Your **${weapon.name}** is broken! Repair it with \`/huntrepair\` or buy a new one with \`/huntbuygun\`.`,
+                content: `Your **${weapon.name}** is broken! Repair it with \`/huntshop repair\` or buy a new one with \`/huntshop weapon\`.`,
                 ephemeral: true
             });
         }
@@ -213,7 +213,7 @@ function buildHuntEmbed(result, user, zone, weapon, currency, discordUser) {
         if (result.expiredCharm) embed.addFields({ name: '🍀 Charm Expired', value: `Your luck charm has worn off.`, inline: false });
 
         if (weapon.status === 'broken') {
-            embed.addFields({ name: '⚠️ Weapon Broke!', value: `Your **${weapon.name}** has broken! Use \`/huntrepair\` before hunting again.`, inline: false });
+            embed.addFields({ name: '⚠️ Weapon Broke!', value: `Your **${weapon.name}** has broken! Use \`/huntshop repair\` before hunting again.`, inline: false });
         } else if (weapon.currentDurability <= Math.floor(weapon.maxDurability * 0.20)) {
             embed.addFields({ name: '⚠️ Low Durability', value: `Your **${weapon.name}** is nearly worn out (${weapon.currentDurability}/${weapon.maxDurability}). Repair soon!`, inline: false });
         }
@@ -250,7 +250,7 @@ function buildHuntEmbed(result, user, zone, weapon, currency, discordUser) {
             embed.addFields({ name: '🛟 Lifesaver Activated!', value: `A severe injury would have destroyed your **${result.deathEvent.weaponName}**, but your Lifesaver absorbed it! (consumed)`, inline: false });
         } else {
             embed.setColor('#8B0000');
-            embed.addFields({ name: '💀 Severe Injury!', value: `The encounter was catastrophic — your **${result.deathEvent.weaponName}** was completely destroyed! Use \`/huntrepair\` to fix it.`, inline: false });
+            embed.addFields({ name: '💀 Severe Injury!', value: `The encounter was catastrophic — your **${result.deathEvent.weaponName}** was completely destroyed! Use \`/huntshop repair\` to fix it.`, inline: false });
         }
     }
 
@@ -260,7 +260,7 @@ function buildHuntEmbed(result, user, zone, weapon, currency, discordUser) {
     }
 
     if (weapon.status === 'broken' && !result.deathEvent) {
-        embed.addFields({ name: '❌ Weapon Broke!', value: `Your **${weapon.name}** has broken! Use \`/huntrepair\` before hunting again.`, inline: false });
+        embed.addFields({ name: '❌ Weapon Broke!', value: `Your **${weapon.name}** has broken! Use \`/huntshop repair\` before hunting again.`, inline: false });
     }
 
     embed.setFooter({ text: 'Tip: Use consumables from /huntshop to boost your success chance' });
