+++ src/commands/utility/help.js
@@ -36,20 +36,16 @@ const CATEGORIES = [
             { name: 'huntinv',        description: 'View and manage your hunt inventory' },
             { name: 'huntprofile',    description: 'View your or another player\'s hunter profile' },
             { name: 'huntquests',     description: 'View and claim your daily hunt quests' },
-            { name: 'huntshop',       description: 'Buy hunting consumables and ammunition' },
-            { name: 'buygun',         description: 'Purchase a weapon for hunting' },
-            { name: 'zone',           description: 'Manage your hunting zones' },
-            { name: 'repair',         description: 'Repair your equipped weapon' },
+            { name: 'huntshop',       description: 'Buy weapons, ammo, consumables, and unlock zones — all in one place' },
+            { name: 'huntzone',       description: 'View and switch your active hunting zone' },
             { name: 'craft',          description: 'Craft items from hunting materials' },
             { name: 'prestige',       description: 'Reset hunter level for permanent prestige bonuses (requires Level 50)' },
             { name: 'fish',           description: 'Cast your line and catch fish (25s cooldown)' },
             { name: 'fishinv',        description: 'View and manage your fishing inventory' },
             { name: 'fishprofile',    description: 'View your or another player\'s fishing profile' },
             { name: 'fishquests',     description: 'View and claim your daily fishing quests' },
-            { name: 'fishshop',       description: 'Browse and purchase fishing supplies' },
-            { name: 'buyrod',         description: 'Purchase a fishing rod or upgrade an existing one' },
-            { name: 'location',       description: 'Manage your fishing locations' },
-            { name: 'fishrepair',     description: 'Repair your equipped fishing rod' },
+            { name: 'fishshop',       description: 'Buy rods, bait, consumables, and unlock locations — all in one place' },
+            { name: 'fishlocation',   description: 'View and switch your active fishing location' },
             { name: 'fishprestige',   description: 'Reset fisher level for permanent prestige bonuses (requires Level 50)' },
             { name: 'blackjack',      description: 'Play blackjack against the dealer' },
             { name: 'baccarat',       description: 'Play a game of baccarat' },
